# src/main/java/com/csebo/fplscraper/fplscraper/api/def/api.yaml
openapi: 3.0.3
info:
  title: Fantasy Premier League Scraper service API - OpenAPI 3.1.0
  description: FPL Scraper API Spec
  version: 1.0.0

paths:
  /data/update/all:
    get:
      tags:
        - data
      summary: "Update stored data by calling FPL server endpoint. Currently stores the following data: players, gameweeks."
      operationId: updateAll
      responses:
        201:
          description: 'Updated players and gameweek deadlines.'
        400:
          description: 'invalid request'
        500:
          description: 'Server side error'

  /league/picks:
    post:
      tags:
        - league
      summary: Returns the players and managers, who selected them in the actual gameweek
      operationId: getPicks
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PicksRequestBody'
      responses:
        200:
          description: 'Returns the picks of managers'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PicksResponseBody'
        400:
          description: 'invalid request'
        500:
          description: 'Server side error'

  /league/transfers:
    post:
      tags:
        - league
      summary: Returns the transfers (in and out) of every manager in the requested gameweek
      operationId: getTransfers
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PicksRequestBody'
      responses:
        200:
          description: 'Returns the transfers of managers in the requested gameweek'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransfersResponseBody'
        400:
          description: 'invalid request'
        500:
          description: 'Server side error'


  /league/data/{leagueId}:
    get:
      tags:
        - league
      summary: Returns basic datas of the league
      operationId: getLeagueData
      parameters:
        - in: path
          description: 'League ID which league to query'
          name: leagueId
          required: true
          schema:
            type: integer
      responses:
        200:
          description: 'Returns basic data of the league'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeagueDataResponseBody'

        400:
          description: 'invalid request'
        500:
          description: 'Server side error'

components:
  schemas:
    PicksRequestBody:
      type: object
      properties:
        playerIds:
          type: array
          items:
            type: integer
        gameweek:
          type: integer
      required:
        - playerIds

    PicksResponseBody:
      description: Returns a map with managers - selected captains entries. Returns a list of players and a list of managers, who selected the actual player.
      type: object
      properties:
        captainPicks:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/CaptainPickDetail'
        picks:
          type: array
          items:
            $ref: '#/components/schemas/PlayerPicksModel'

    TransfersResponseBody:
      description: Returns a the in-out transfers of each manager.
      type: object
      properties:
        transfers:
          type: object
          additionalProperties:
            type: array
            items:
              $ref: '#/components/schemas/TransferredPlayerModel'

    PlayerPicksModel:
      description: Returns a list of players and a list of managers, who selected the actual player.
      type: object
      properties:
        playerId:
          description: 'ID of the football player'
          type: integer
        playerName:
          description: 'Name of the football player'
          type: string
        numberOfPicks:
          description: 'How many times the football player were picked'
          type: integer
        pickedBy:
          type: array
          items:
            $ref: '#/components/schemas/ManagerPickModel'

    TransferredPlayerModel:
      type: object
      properties:
        playerName:
          type: string
        in:
          type: boolean
      required:
        - playerName
        - in

    ManagerPickModel:
      description: 'Represents a single pick of a football player by a manager.'
      type: object
      properties:
        id:
          description: 'ID of the manager'
          type: integer
        multiplier:
          description: 'Multiplier of the player'
          type: integer
      required:
        - id
        - name
        - multiplier

    CaptainPickDetail:
      type: object
      properties:
        playerName:
          type: string
        multiplier:
          type: integer
      required:
        - playerName
        - multiplier

    LeagueDataResponseBody:
      description: 'Contains basic data of the league'
      type: object
      properties:
        leagueName:
          description: 'Name of the league'
          type: string
        managers:
          additionalProperties:
            type: string